Add Windsor support
An attempt to take the AutoFac integration and convert it
to Windsor support. One key difficulty is a lack of
understanding how to translate AutoFac LifetimeScope to
something that is idiomatic to Windsor which doesn't like
the concept of Child Containers and prefers alternate
methods for acheiving the same thing.